Can supply own config file, added multiple other CLI options.
Can supply custom config file with the config option on launch:
	ruby cosa.rb --config, -c myconfig.yaml

Can change the output method with one of the output methods on launch:
	ruby cosa.rb --silent, -s for silent output
	ruby cosa.rb --verbose, -v for verbose output

Can clear the queue with the clear queue option:
	ruby cosa.rb --clear-queue, -e

Can add a new url to the queue:
	ruby cosa.rb --add, -a http://wwww.example.com

Can list the css urls that are linked to:
	ruby cosa.rb --css

Can list the urls that link to a given url:
	ruby cosa.rb --to, -t http://www.example.com

Can list the urls that a given url link to:
	ruby cosa.rb --from, -f http://www.example.com

Can list the urls that took longer than a given time to respond:
	ruby cosa.rb --response-time, -r 0.2

Also started work on the create_config function that will be called if a config is not present or given.

Fixed some errors with the SHELF time not being retrieved.

When run with no given url, the crawler will no longer recrawl the site.