Minor changes to banner text/help message. Added validation to some of the command line options. More additions to create_config.
Cosa will now check to make sure that the values given for --age and --config are valid.
If a date is given for --age that is not in the form of yyyy-mm-dd, or
if the value given for --config does not end in .yaml then Cosa will throw an error.