Additions to create_config, can now create new databases, removed old require statements.
create_config method will now:
	check if the given path exists.
	check if the given file exists and call create_db if it does not.

create_db creates a connection to an sqlite database at the given path with the given filename.
Then creates the necessary tables for the crawler to run.

Still need to make changes to this, as the crawler runs slower when inserting into a
dynamically created table.