Added the info option.
The info option allows you to get any information on a url that was stored in the meta table.

For example:
cosa -I http://www.example.com/image.jpg
would output:
http://www.example.com/image.jpg
dimension-x: 800
dimension-y: 600

Currently the only items being added to the meta table are images, and the only information
being stored are width and height.